Now correctly adds the json to a specified file

+++ include/ParseEventProperties.cpp
@@ -11,10 +11,10 @@ std::vector<Topic> ParseEventProperties(std::vector<soap_dom_element> dom) {
 
 	std::stringstream stream;
 	//Parse
-	for (int i = 0; i < dom.size(); i++) {
+	for (size_t i = 0; i < dom.size(); i++) {
 
 		//This clearly is not the most effiecient way of doing this, but soap_dom does not provide a distance class
-		int depth = 0;
+		size_t depth = 0;
 		for (soap_dom_element::iterator it = dom[i].begin(); it != dom[i].end(); ++it) {
 			if (it->depth() > depth) { depth = it->depth(); }
 		}
@@ -29,7 +29,7 @@ std::vector<Topic> ParseEventProperties(std::vector<soap_dom_element> dom) {
 					soap_dom_attribute dom = *it->atts;
 					soap_dom_attribute domNext = *it->atts->next;
 
-					for (int j = 0; j < depth; j++) {
+					for (size_t j = 0; j < depth; j++) {
 						if (Topic[j].find("tt:") == std::string::npos) {
 							//Make look better
 							if (Topic[j].find(':') != std::string::npos) { stream << split(Topic[j], ':')[1] << "/"; }
@@ -54,10 +54,10 @@ std::vector<Topic> ParseEventProperties(std::vector<soap_dom_element> dom) {
 	std::vector<std::string> splitByLine = split(stream.str(), '\n');
 	std::vector<std::vector<std::string>> splitBySpace;
 
-	for (int i = 0; i < splitByLine.size(); i++) {
+	for (size_t i = 0; i < splitByLine.size(); i++) {
 		std::vector<std::string> vec1 = split(splitByLine[i], ' ');
 		bool boo = true;
-		for (int j = 0; j < splitBySpace.size(); j++) {
+		for (size_t j = 0; j < splitBySpace.size(); j++) {
 			if (splitBySpace[j][0] != vec1[0]) {
 				boo = true;
 			}
@@ -79,14 +79,14 @@ std::vector<Topic> ParseEventProperties(std::vector<soap_dom_element> dom) {
 	}
 
 	std::vector<Topic> topics;
-	for (int i = 0; i < splitBySpace.size(); i++) {
+	for (size_t i = 0; i < splitBySpace.size(); i++) {
 
 		Topic topic;
 		topic.name = splitBySpace[i][0];
 
 		std::vector<std::pair<std::string, std::string>> elements;
 
-		for (int k = 1; k < splitBySpace[i].size(); k += 2) {
+		for (size_t k = 1; k < splitBySpace[i].size(); k += 2) {
 			std::pair<std::string, std::string> pair = make_pair(splitBySpace[i][k], splitBySpace[i][k + 1]);
 			elements.push_back(pair);
 		}
